requirements.txt
PySide6
sounddevice
keyboard
requests
configparser
pytest 
pygame
python-dotenv
soundfile
requests
assemblyai
google-generativeai